requirements.txt
python-telegram-bot==20.8
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
